land_cover_classification_backend/requirements.txt
albumentations==1.3.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
efficientnet-pytorch==0.7.1
filelock==3.12.0
firebase-admin==6.1.0
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.39.3
fsspec==2023.4.0
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
httplib2==0.22.0
huggingface-hub==0.14.0
idna==3.4
imageio==2.28.0
importlib-metadata==6.3.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
lit==16.0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
msgpack==1.0.5
munch==2.5.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pretrainedmodels==0.7.4
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.2
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
segmentation-models-pytorch==0.3.2
shapely==2.0.1
six==1.16.0
sympy==1.11.1
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.6.12
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
